trendsLegendDict: move to dict folder
