+++ src/dict/trendsLegendDict.js
